add comments in aaz._command
