+++ src/azure-cli-core/azure/cli/core/aaz/_utils.py
@@ -10,14 +10,6 @@ from azure.cli.core.aaz.exceptions import AAZInvalidShorthandSyntaxError
 from ._help import AAZShowHelp
 
 
-def _get_profile_pkg(aaz_module_name, cloud):
-    profile_module_name = cloud.profile.lower().replace('-', '_')
-    try:
-        return importlib.import_module(f'{aaz_module_name}.{profile_module_name}')
-    except ModuleNotFoundError:
-        return None
-
-
 class AAZShortHandSyntaxParser:
     NULL_EXPRESSIONS = ('null', 'None')
     HELP_EXPRESSIONS = ('??', )
